requirements.txt
naff
python-dotenv
orjson
samp-client